refactor: Ping과 Chat 타입 별 provider 생성
